% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_portfolio_weights}
\alias{demo_portfolio_weights}
\title{Demo portfolio weights}
\format{
A data frame with \code{asset} and \code{weight}.
}
\source{
Created for package examples.
}
\usage{
demo_portfolio_weights
}
\description{
Example equity, bond, and cash portfolio weights.
}
\keyword{datasets}
