% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_return_data}
\alias{demo_return_data}
\title{Demo return data}
\format{
A data frame with \code{date}, \code{equity_return}, \code{bond_return}, and \code{cash_return}.
}
\source{
Prepared for package examples.
}
\usage{
demo_return_data
}
\description{
Prepared monthly asset return data used in package examples.
}
\keyword{datasets}
